+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.1
+## 1.0.0 - 2025-06-06
 
-* TODO: Describe initial release.
+- :) Initial release
+- Added InfiniteAgeScroller widget
+- Supports infinite vertical, horizontal age scrolling
+- Supports infinite left diagonal, right diagonal age scrolling
+- Customizable callback for age selection
+- Includes example app
+- To be Continued